Game will now spawn a variable number of tracks and track managers that run separate tracks

+++ Source/Endless_Runner/Managers/TrackManager.h
@@ -52,7 +52,7 @@ public:
 	UPROPERTY(BlueprintReadOnly)
 	TArray<ATrackPiece*> CurrentTrackPieces;
 	UPROPERTY(EditAnywhere)
-	float TrackLength;
+	int TrackLength;
 	UPROPERTY(EditAnywhere) 
 	float TrackSpeed;
